requirements_postgres.txt
# Updated Patient Management System Dependencies
# Compatible with Python 3.8+ - PostgreSQL Version

# Core Web Framework
fastapi==0.104.1                # Modern, fast web framework for building APIs
uvicorn[standard]==0.24.0        # Lightning-fast ASGI server implementation

# Database and ORM
sqlalchemy==2.0.23              # SQL toolkit and Object-Relational Mapping (ORM)
psycopg2-binary==2.9.9          # PostgreSQL adapter for Python
alembic==1.12.1                 # Database migration tool for SQLAlchemy

# Data Validation and Serialization  
pydantic[email]==2.5.0           # Data validation with email support
email-validator==2.1.0          # Email validation for Pydantic

# Form Handling
python-multipart==0.0.6         # Multipart form data parsing

# Authentication and Security - HIPAA Enhanced
python-jose[cryptography]==3.3.0  # JWT token handling
passlib[bcrypt]==1.7.4            # Password hashing with bcrypt
cryptography==41.0.7             # Encryption for PHI data
python-dotenv==1.0.0             # Environment variable management
bcrypt==4.0.1                     # Fixed bcrypt version for compatibility

# Session Management
itsdangerous==2.1.2               # Secure session cookies

# PostgreSQL Connection Pooling
psycopg2-pool==1.1               # Advanced connection pooling
sqlalchemy-utils==0.41.1         # SQLAlchemy utility functions

# Database Migration and Backup
pg_dump==1.0.0                   # PostgreSQL backup utilities
sqlparse==0.4.4                  # SQL parsing for migrations

# Development and Testing Tools
pytest==7.4.3                    # Testing framework
pytest-asyncio==0.21.1           # Async testing support
httpx==0.25.2                    # HTTP client for testing APIs

# Monitoring and Logging
psutil==5.9.6                    # System and process monitoring
structlog==23.2.0                # Structured logging
